+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres
 
+## [0.1.5] - 2025-08-31
+### Added
+- **BBMM:** added minimal v12 support. In v12 just added tools to export settings and module states to .json to prep for migration to v13.
+    - NOTE! Cannot guarantee all settings will import in v13 if setting names have changed. 
+
 ## [0.1.4] - 2025-08-30
 ### Fixed
 - **Module Preset Manager:** Fixed dual 'export to .json' buttons.